upstash_Vector/requirements.txt

langchain-community==0.2.16
langchain-text-splitters==0.2.4
lanchain==
langchain-groq==0.1.9
unstructured==0.15.0
unstructured[pdf]==0.15.0
nltk==3.8.1
upstash-vector
python-dotenv

lanchain